Use 'resources' instead of 'sources'
According to the buster-configuration documentation for sources (at "http://docs.busterjs.org/en/latest/modules/buster-configuration/"): 

"sources: Source files to load in script tags in the browser." 

But script tags duplicate what requirejs already does to load source in tests whenever it encounters a "define()" block.  So, while using 'sources' works, it's a bit of a misdirection in terms of communicating intent.  I offer that it's better to use the 'resources' field, which seems more specific to this use case.  I've updated the configuration and tested that the tests run as before.